Added user Authentication to requests

+++ server/node/server.js
@@ -63,8 +63,28 @@ async function query(q, params) {
 	return query;
 }
 
+function fromB64(str) {
+    return Buffer.from(str, "base64").toString("utf8");
+}
+
+function isString(obj) {
+	return obj && typeof obj === 'string' || obj instanceof String;
+}
+
+function asNumber(obj) {
+    if (obj) {
+        var n = Number(obj);
+
+        if (n !== NaN) {
+            return n;
+        }
+    }
+
+    return null;
+}
+
 function convertUserToEmail(user) {
-    if (user && isString(user)) {
+    if (isString(user)) {
         var u = user;
 
         if (u.toLowerCase().indexOf("k") === -1) {
@@ -81,7 +101,7 @@ function convertUserToEmail(user) {
 }
 
 function convertEmailToUser(user) {
-    if (user && isString(user)) {
+    if (isString(user)) {
         var u = user;
         if (u.indexOf("@") !== -1) {
             u = u.split("@")[0];
@@ -101,31 +121,35 @@ function convertEmailToUser(user) {
 }
 
 async function isUserAuthorisedForWrite(user) {
-    const q = await query("SELECT * FROM authorised_logins WHERE `email` = ? ", [convertUserToEmail(user)]);
+    if (isString(user)) {
+        if (user.indexOf("Basic") !== -1) {
+            user = fromB64(user.split(" ")[1]);
+            if (user.indexOf(":") !== -1) {
+                user = user.replace(":", "");
+            }
 
-    return q && q[0][0].length > 0;
+            const q = await query("SELECT * FROM authorised_logins WHERE `email` = ? ", [convertUserToEmail(user)]);
+
+            return q && q[0].length > 0;
+        }
+    }
+    return false;
 }
 
 async function isUserAuthorisedForRead(user) {
-    const q = await query("SELECT * FROM students_in_groups WHERE `k_number` = ?", [convertEmailToUser(user)]);
+    if (isString(user)) {
+        if (user.indexOf("Basic") !== -1) {
+            user = fromB64(user.split(" ")[1]);
+            if (user.indexOf(":") !== -1) {
+                user = user.replace(":", "");
+            }
 
-    return q && (q[0][0].length > 0 || isUserAuthorisedForWrite(user));
-}
+            const q = await query("SELECT * FROM students_in_groups WHERE `k_number` = ?", [convertEmailToUser(user)]);
 
-function isString(obj) {
-	return obj && typeof obj === 'string' || myVar instanceof String;
-}
-
-function asNumber(obj) {
-    if (obj) {
-        var n = Number(obj);
-
-        if (n !== NaN) {
-            return n;
+            return q && (q[0].length > 0 || isUserAuthorisedForWrite(user));
         }
     }
-
-    return null;
+    return false;
 }
 
 async function getGroupByNameOrId(gid, gname) {
@@ -164,23 +188,32 @@ async function getGroupByNameOrId(gid, gname) {
 
 router.get("/api/v1.0/group/", upload.array(), asyncMiddleware(async (req, res, next) => {
 	if (req.secure && req.accepts('application/json')) {
-        var groupq = await getGroupByNameOrId(req.query.groupId, req.query.groupName);
+        if (await isUserAuthorisedForRead(req.get("authorization"))) {
+            var groupq = await getGroupByNameOrId(req.query.groupId, req.query.groupName);
 
-        if (groupq.length > 0) {
-            res.status(200).json({"status": 200, "message": "Ok: group found", "groupId": groupq[0][0].module_id, "groupName": groupq[0][0].module_name});
-        } else if (req.query.groupId || req.query.groupName) {
-            res.status(400).json({"status": 400, "message": "Bad Request: groupId or groupName isn't valid"});
-        } else {
-            const q = await query("SELECT * FROM modules");
+            if (groupq.length > 0) {
+                res.status(200).json({"status": 200, "message": "Ok: group found", "groupId": groupq[0][0].module_id, "groupName": groupq[0][0].module_name});
+            } else if (req.query.groupId || req.query.groupName) {
+                res.status(400).json({"status": 400, "message": "Bad Request: groupId or groupName isn't valid"});
+            } else {
+                user = fromB64(req.get("authorization").split(" ")[1]);
+                if (user.indexOf(":") !== -1) {
+                    user = user.replace(":", "");
+                }
 
-            var groups = [];
+                const q = await query("SELECT * FROM modules WHERE `module_id` in (SELECT module_id FROM students_in_groups WHERE `k_number` = ?)", [convertEmailToUser(user)]);
 
-            for (var i in q[0]) {
-                var row = q[0][i];
-                groups.push({"groupId": row.module_id, "groupName": row.module_name});
+                var groups = [];
+
+                for (var i in q[0]) {
+                    var row = q[0][i];
+                    groups.push({"groupId": row.module_id, "groupName": row.module_name});
+                }
+
+                res.status(200).json({"status": 200, "message": "Ok: All " + q[0].length + " group(s) have been selected", "groups": groups});
             }
-
-            res.status(200).json({"status": 200, "message": "Ok: All " + q[0].length + " group(s) have been selected", "groups": groups});
+        } else {
+            res.status(401).json({"status": 401, "message": "Unauthorised"});
         }
 	} else {
 		res.sendStatus(406);
@@ -189,19 +222,22 @@ router.get("/api/v1.0/group/", upload.array(), asyncMiddleware(async (req, res, 
 
 router.post("/api/v1.0/group/", upload.array(), asyncMiddleware(async (req, res, next) => {
 	if (req.secure && req.accepts('application/json')) {
-		if (req.body.groupName && isString(req.body.groupName)) {
-            console.log("SELECT * FROM modules WHERE `module_name` = " + req.body.groupName);
-			const existing = await query("SELECT * FROM modules WHERE `module_name` = ?", [req.body.groupName]);
-			if (existing[0].length == 0) {
-				var group = { module_name: req.body.groupName };
-				const q = await query('INSERT INTO modules SET ?', group);
-				res.status(201).json({"status": 201, "message": "Created: group created", "groupid": q[0].insertId, "groupName": req.body.groupName});
-			} else {
-				res.status(400).json({"status": 400, "message": "Bad Request: Group already exists. If you want information about the group use GET, or if you want to add/remove members use PUT"});
-			}
-		} else {
-			res.status(400).json({"status": 400, "message": "Bad Request: groupName must be specified in the body"});
-		}
+        if (await isUserAuthorisedForWrite(req.get("authorization"))) {
+    		if (req.body.groupName && isString(req.body.groupName)) {
+    			const existing = await query("SELECT * FROM modules WHERE `module_name` = ?", [req.body.groupName]);
+    			if (existing[0].length == 0) {
+    				var group = { module_name: req.body.groupName };
+    				const q = await query('INSERT INTO modules SET ?', group);
+    				res.status(201).json({"status": 201, "message": "Created: group created", "groupid": q[0].insertId, "groupName": req.body.groupName});
+    			} else {
+    				res.status(400).json({"status": 400, "message": "Bad Request: Group already exists. If you want information about the group use GET, or if you want to add/remove members use PUT"});
+    			}
+    		} else {
+    			res.status(400).json({"status": 400, "message": "Bad Request: groupName must be specified in the body"});
+    		}
+        } else {
+            res.status(401).json({"status": 401, "message": "Unauthorised"});
+        }
 	} else {
 		res.sendStatus(406);
 	}
@@ -209,50 +245,54 @@ router.post("/api/v1.0/group/", upload.array(), asyncMiddleware(async (req, res,
 
 router.put("/api/v1.0/group/", upload.array(), asyncMiddleware(async (req, res, next) => {
 	if (req.secure && req.accepts('application/json')) {
-		if (req.body.users && req.body.users.constructor === Array) {
-            var groupq = await getGroupByNameOrId(req.body.groupId, req.body.groupName);
-            if (groupq.length > 0) {
-                var groupid = groupq[0][0].module_id,
-                    cleanUsers = [],
-                    users = req.body.users;
+        if (await isUserAuthorisedForWrite(req.get("authorization"))) {
+    		if (req.body.users && req.body.users.constructor === Array) {
+                var groupq = await getGroupByNameOrId(req.body.groupId, req.body.groupName);
+                if (groupq.length > 0) {
+                    var groupid = groupq[0][0].module_id,
+                        cleanUsers = [],
+                        users = req.body.users;
 
-                for (var i in users) {
-                    var user = users[i];
-                    if (user && isString(user)) {
-                        if (user.indexOf("@") !== -1) {
-                            user = user.split("@")[0];
-                        }
+                    for (var i in users) {
+                        var user = users[i];
+                        if (user && isString(user)) {
+                            if (user.indexOf("@") !== -1) {
+                                user = user.split("@")[0];
+                            }
 
-                        if (user.toLowerCase().indexOf("k") !== -1) {
-                            user = user.toLowerCase().replace("k", "");
-                        }
+                            if (user.toLowerCase().indexOf("k") !== -1) {
+                                user = user.toLowerCase().replace("k", "");
+                            }
 
-                        user = Number(user);
+                            user = Number(user);
 
-                        if (user !== NaN) {
-                            cleanUsers.push(user);
+                            if (user !== NaN) {
+                                cleanUsers.push(user);
+                            }
                         }
                     }
-                }
 
-                var createdCount = 0;
+                    var createdCount = 0;
 
-                for (var i in cleanUsers) {
-                    var user = cleanUsers[i];
-                    const sel = await query("SELECT * FROM students_in_groups WHERE `module_id` = ? and `k_number` = ?", [groupid, user]);
-                    if (sel[0].length == 0) {
-                        const q = await query("INSERT INTO students_in_groups SET ?", {"module_id": groupid, "k_number": user});
-                        createdCount += 1;
+                    for (var i in cleanUsers) {
+                        var user = cleanUsers[i];
+                        const sel = await query("SELECT * FROM students_in_groups WHERE `module_id` = ? and `k_number` = ?", [groupid, user]);
+                        if (sel[0].length == 0) {
+                            const q = await query("INSERT INTO students_in_groups SET ?", {"module_id": groupid, "k_number": user});
+                            createdCount += 1;
+                        }
                     }
-                }
 
-                res.status(201).json({"status": 201, "message": "Created: " + createdCount + " user(s) added to the group " + groupid});
-            } else {
-                res.status(400).json({"status": 400, "message": "Bad Request: groupId or groupName must be specified in the body"});
-            }
-		} else {
-			res.status(400).json({"status": 400, "message": "Bad Request: users parameter must be specified as an array in the body"});
-		}
+                    res.status(201).json({"status": 201, "message": "Created: " + createdCount + " user(s) added to the group " + groupid});
+                } else {
+                    res.status(400).json({"status": 400, "message": "Bad Request: groupId or groupName must be specified in the body"});
+                }
+    		} else {
+    			res.status(400).json({"status": 400, "message": "Bad Request: users parameter must be specified as an array in the body"});
+    		}
+        } else {
+            res.status(401).json({"status": 401, "message": "Unauthorised"});
+        }
 	} else {
 		res.sendStatus(406);
 	}
@@ -260,15 +300,19 @@ router.put("/api/v1.0/group/", upload.array(), asyncMiddleware(async (req, res, 
 
 router.delete("/api/v1.0/group/", upload.array(), asyncMiddleware(async (req, res, next) => {
 	if (req.secure && req.accepts('application/json')) {
-        var groupq = await getGroupByNameOrId(req.body.groupId, req.body.groupName);
-        if (groupq.length > 0) {
-            var groupid = groupq[0][0].module_id;
-			const q = await query("DELETE FROM modules WHERE `module_id` = ?", [groupid]);
+        if (await isUserAuthorisedForWrite(req.get("authorization"))) {
+            var groupq = await getGroupByNameOrId(req.body.groupId, req.body.groupName);
+            if (groupq.length > 0) {
+                var groupid = groupq[0][0].module_id;
+    			const q = await query("DELETE FROM modules WHERE `module_id` = ?", [groupid]);
 
-			res.status(200).json({"status": 200, "message": "Ok: " + q[0].affectedRows + " group(s) deleted"});
-		} else {
-			res.status(400).json({"status": 400, "message": "Bad Request: groupName or groupId must be specified in the body"});
-		}
+    			res.status(200).json({"status": 200, "message": "Ok: " + q[0].affectedRows + " group(s) deleted"});
+    		} else {
+    			res.status(400).json({"status": 400, "message": "Bad Request: groupName or groupId must be specified in the body"});
+    		}
+        } else {
+            res.status(401).json({"status": 401, "message": "Unauthorised"});
+        }
 	} else {
 		res.sendStatus(406);
 	}
@@ -276,6 +320,7 @@ router.delete("/api/v1.0/group/", upload.array(), asyncMiddleware(async (req, re
 
 router.get("/api/v1.0/announcement/:groupid", upload.array(), asyncMiddleware(async (req, res, next) => {
     if (req.secure && req.accepts('application/json')) {
+        if (await isUserAuthorisedForRead(req.get("authorization"))) {
             var retval;
             if (req.query) {
                 const time = asNumber(req.query.date);
@@ -299,6 +344,9 @@ router.get("/api/v1.0/announcement/:groupid", upload.array(), asyncMiddleware(as
             } else {
                 res.status(400).json({"status": 400, "message": "Bad Request: no parameters"});
             }
+        } else {
+            res.status(401).json({"status": 401, "message": "Unauthorised"});
+        }
     } else {
         res.sendStatus(406);
     }
@@ -306,16 +354,20 @@ router.get("/api/v1.0/announcement/:groupid", upload.array(), asyncMiddleware(as
 
 router.delete("/api/v1.0/announcement/", upload.array(), asyncMiddleware(async (req, res, next) => {
     if (req.secure && req.accepts('application/json')) {
-        if (req.body.announcement_id) {
-            var announcementExists = await query('SELECT * FROM messages_sent WHERE `announcement_id` = ?', [req.body.announcement_id]);
-            if (!announcementExists[0].length) {
-                res.status(400).json({"status": 400, "message": "Bad Request: announcement does not exist."});
+        if (await isUserAuthorisedForWrite(req.get("authorization"))) {
+            if (req.body.announcement_id) {
+                var announcementExists = await query('SELECT * FROM messages_sent WHERE `announcement_id` = ?', [req.body.announcement_id]);
+                if (!announcementExists[0].length) {
+                    res.status(400).json({"status": 400, "message": "Bad Request: announcement does not exist."});
+                } else {
+                    const deleteQuery = await query('DELETE FROM messages_sent WHERE `announcement_id` = ?', [req.body.announcement_id]);
+                    res.status(202).json({"status": 202, "message": "Successfully deleted announcement."});
+                }
             } else {
-                const deleteQuery = await query('DELETE FROM messages_sent WHERE `announcement_id` = ?', [req.body.announcement_id]);
-                res.status(202).json({"status": 202, "message": "Successfully deleted announcement."});
+                res.status(400).json({"status": 400, "message": "Bad Request: announcement id missing."});
             }
         } else {
-            res.status(400).json({"status": 400, "message": "Bad Request: announcement id missing."});
+            res.status(401).json({"status": 401, "message": "Unauthorised"});
         }
     } else {
         res.sendStatus(406);
@@ -324,30 +376,34 @@ router.delete("/api/v1.0/announcement/", upload.array(), asyncMiddleware(async (
 
 router.post("/api/v1.0/announcement/", upload.array(), asyncMiddleware(async (req, res, next) => {
     if (req.secure && req.accepts('application/json')) {
-        var groupq = await getGroupByNameOrId(req.body.groupId, req.body.groupName);
-        if (groupq.length > 0 && isString(req.body.announcement) && isString(req.body.subject)) {
-            var groupid = groupq[0][0].module_id;
-            var insert = {module_id: groupid, message: req.body.announcement, subject: req.body.subject};
-            const q = await query('INSERT INTO messages_sent SET ?', insert);
-            var message = {
-                android: {
-                    priority: "high"
-                },
-                data: {
-                    subject: req.body.subject,
-                    message: req.body.announcement,
-                    group: groupid.toString()
-                },
-                condition: "'announcements' in topics"
-            };
+        if (await isUserAuthorisedForWrite(req.get("authorization"))) {
+            var groupq = await getGroupByNameOrId(req.body.groupId, req.body.groupName);
+            if (groupq.length > 0 && isString(req.body.announcement) && isString(req.body.subject)) {
+                var groupid = groupq[0][0].module_id;
+                var insert = {module_id: groupid, message: req.body.announcement, subject: req.body.subject};
+                const q = await query('INSERT INTO messages_sent SET ?', insert);
+                var message = {
+                    android: {
+                        priority: "high"
+                    },
+                    data: {
+                        subject: req.body.subject,
+                        message: req.body.announcement,
+                        group: groupid.toString()
+                    },
+                    condition: "'announcements' in topics"
+                };
 
-            admin.messaging().send(message).then((response) => {
-                res.status(201).json({"status": 201, "message": "Created: successfully sent new announcement", "response": response});
-            }).catch((error) => {
-                console.log('Error sending message:', error);
-            });
+                admin.messaging().send(message).then((response) => {
+                    res.status(201).json({"status": 201, "message": "Created: successfully sent new announcement", "response": response});
+                }).catch((error) => {
+                    console.log('Error sending message:', error);
+                });
+            } else {
+                res.status(400).json({"status": 400, "message": "Bad Request: groupName or groupId, announcement and subject parameters must be specified in the body"});
+            }
         } else {
-            res.status(400).json({"status": 400, "message": "Bad Request: groupName or groupId, announcement and subject parameters must be specified in the body"});
+            res.status(401).json({"status": 401, "message": "Unauthorised"});
         }
     } else {
         res.sendStatus(406);
